admin API add feature

+++ project/Backend/routes/admin.js
@@ -7,25 +7,32 @@ const isNumeric = (value) => /^-?\d+$/.test(value);
 
 router.get('/unconfirmed', async (req, res) => {
 
-    const { rows: comments } = await db.query("SELECT id, comment FROM comments "
-        + "WHERE confirmed = FALSE", [])
-    
+    const { rows: comments } = await db.query("SELECT id, comment FROM comments WHERE confirmed = FALSE")
+
     return res.status(200).send({ comments })
 });
 
 router.delete('/feedback/:id', async (req, res) => {
     const { id: commentId } = req.params
+    
+    if (!isNumeric(commentId))
+        return res.status(400).send({ message: 'comment id is not valid' })
 
     await db.query("DELETE FROM comments WHERE id = $1", [commentId]);
 
     return res.sendStatus(204);
 });
 
-router.put('/confirm/:id', async (req, res) => {
+router.put('/confirm/:id?', async (req, res) => {
     const { id: commentId } = req.params
 
-    await db.query("UPDATE comments SET confirmed = TRUE WHERE id = $1", [commentId]);
-
+    if (commentId === undefined) {
+        await db.query("UPDATE comments SET confirmed = TRUE");
+    } else {
+        if (!isNumeric(commentId))
+            return res.status(400).send({ message: 'comment id is not valid' })
+        await db.query("UPDATE comments SET confirmed = TRUE WHERE id = $1", [commentId]);
+    }
     return res.sendStatus(204);
 });
